Проектная работа 14

+++ controllers/cards.js
@@ -1,58 +1,54 @@
 const mongoose = require('mongoose');
 const Card = require('../models/card');
 const Constants = require('../utils/constants');
+const NotFoundError = require('../middlewares/errors/not-found-err');
+const BadRequestError = require('../middlewares/errors/bad-request');
+const OwnerError = require('../middlewares/errors/owner-error');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(Constants.INTERNAL_SERVER_ERROR).send({
-      message: Constants.SERVER_ERROR,
-    }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err instanceof mongoose.Error.ValidationError) {
-        res.status(Constants.BAD_REQUEST).send({
-          message: Constants.CREATE_CARD_INCORRECT_DATA,
-        });
+        next(new BadRequestError(Constants.CREATE_CARD_INCORRECT_DATA));
       } else {
-        res.status(Constants.INTERNAL_SERVER_ERROR).send({
-          message: Constants.SERVER_ERROR,
-        });
+        next(err);
       }
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => {
-      if (card) {
-        res.send({ data: card });
-      } else {
-        res.status(Constants.NOT_FOUND).send({
-          message: Constants.PASSED_NON_EXISTENT_CARD_ID,
+module.exports.deleteCard = async (req, res, next) => {
+  try {
+    const cardId = await Card.findOne({ _id: req.params.cardId });
+    const cardOwner = req.user._id;
+    if (cardId === null) {
+      next(new NotFoundError(Constants.NOT_FOUND_CARD_WITH_ID));
+    } else if (cardId.owner.valueOf() === cardOwner) {
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((card) => {
+          res.send({ data: card });
         });
-      }
-    })
-    .catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        res.status(Constants.BAD_REQUEST).send({
-          message: Constants.INVALID_CARD_ID,
-        });
-      } else {
-        res.status(Constants.INTERNAL_SERVER_ERROR).send({
-          message: Constants.SERVER_ERROR,
-        });
-      }
-    });
+    } else {
+      next(new OwnerError(Constants.DELETE_PROHIBITED));
+    }
+  } catch (err) {
+    if (err instanceof mongoose.Error.CastError) {
+      next(new BadRequestError(Constants.INVALID_CARD_ID));
+    } else {
+      next(err);
+    }
+  }
 };
 
-module.exports.likeCard = (req, res) => Card.findByIdAndUpdate(
+module.exports.likeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $addToSet: { likes: req.user._id } },
   { new: true },
@@ -61,24 +57,18 @@ module.exports.likeCard = (req, res) => Card.findByIdAndUpdate(
     if (card) {
       res.send({ data: card });
     } else {
-      res.status(Constants.NOT_FOUND).send({
-        message: Constants.NOT_FOUND_CARD_WITH_ID,
-      });
+      next(new NotFoundError(Constants.NOT_FOUND_CARD_WITH_ID));
     }
   })
   .catch((err) => {
     if (err instanceof mongoose.Error.CastError) {
-      res.status(Constants.BAD_REQUEST).send({
-        message: Constants.LIKE_OR_DISLIKE_INCORRECT_DATA,
-      });
+      next(new BadRequestError(Constants.LIKE_OR_DISLIKE_INCORRECT_DATA));
     } else {
-      res.status(Constants.INTERNAL_SERVER_ERROR).send({
-        message: Constants.SERVER_ERROR,
-      });
+      next(err);
     }
   });
 
-module.exports.dislikeCard = (req, res) => Card.findByIdAndUpdate(
+module.exports.dislikeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $pull: { likes: req.user._id } },
   { new: true },
@@ -87,18 +77,13 @@ module.exports.dislikeCard = (req, res) => Card.findByIdAndUpdate(
     if (card) {
       res.send({ data: card });
     } else {
-      res.status(Constants.NOT_FOUND).send({
-        message: Constants.NOT_FOUND_CARD_WITH_ID,
-      });
+      next(new NotFoundError(Constants.NOT_FOUND_CARD_WITH_ID));
     }
   })
   .catch((err) => {
     if (err instanceof mongoose.Error.CastError) {
-      return res.status(Constants.BAD_REQUEST).send({
-        message: Constants.LIKE_OR_DISLIKE_INCORRECT_DATA,
-      });
+      next(new BadRequestError(Constants.LIKE_OR_DISLIKE_INCORRECT_DATA));
+    } else {
+      next(err);
     }
-    return res.status(Constants.INTERNAL_SERVER_ERROR).send({
-      message: Constants.SERVER_ERROR,
-    });
   });
